requirements.txt
Flask
psycopg2
SQLAlchemy
Flask-SQLAlchemy
python-dotenv